New Features 🎉
Schema can be shown on the page.
Links works with Launches!
Links can be copied and open in the next tab.
Templates instances can be outlined with a blue border.
MSK/QA Flag on localStorage can be now toggle.

Spacers are now showed on the page, but only works on 1280 breakpoint.